move sample tsv creation code to sample.py
