+++ rotary/rotary.py
@@ -6,7 +6,6 @@ Created by: Lee Bergstrand (2023)
 Description: A command-line tool for the Rotary hybrid assembly workflow.
 """
 import argparse
-import csv
 import os
 import subprocess
 import sys
@@ -14,7 +13,7 @@ import sys
 import psutil
 from ruamel.yaml import YAML
 
-from rotary.sample import SequencingFile, Sample, is_fastq_file
+from rotary.sample import SequencingFile, Sample, is_fastq_file, create_sample_tsv
 from rotary.utils import check_for_files, get_cli_arg_path
 
 yaml = YAML()
@@ -233,22 +232,6 @@ def init(args, config, output_dir_path):
     write_config_file(config, output_dir_path)
 
 
-def create_sample_tsv(output_dir_path, samples):
-    """
-    Generates a TSV file in the output directory with a series of CLI paths for files belonging to each sample.
-
-    :param output_dir_path: The path to the output Rotary directory.
-    :param samples: A list of Sample objects.
-    """
-    sample_tsv_path = os.path.join(output_dir_path, 'samples.tsv')
-    with open(sample_tsv_path, 'w') as tsv_file:
-        tsv_writer = csv.writer(tsv_file, delimiter='\t')
-        header = ['sample_id', 'long-read', 'short-read_R1', 'short-read_R2']
-        tsv_writer.writerow(header)
-        for current_sample in samples:
-            tsv_writer.writerow(current_sample.sample_file_row)
-
-    return sample_tsv_path
 
 
 def write_config_file(config, output_dir_path):
